% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_store.R
\name{wkfeaturestore_command}
\alias{wkfeaturestore_command}
\title{Apply a command JSON and return the resulting response JSON.}
\usage{
wkfeaturestore_command(store, cmd_json)
}
\arguments{
\item{store}{A store handle from [wkfeaturestore_new()].}

\item{cmd_json}{A command JSON string.}
}
\value{
The response as a JSON string.
}
\description{
Apply a command JSON and return the resulting response JSON.
}
