% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_store.R
\name{wkfeaturestore_version}
\alias{wkfeaturestore_version}
\title{The wickra-feature-store library version.}
\usage{
wkfeaturestore_version()
}
\value{
A version string.
}
\description{
The wickra-feature-store library version.
}
